+++ demo/filescreen.cpp
@@ -12,7 +12,7 @@ void fd::FileScreen::reset (GameState* gs) {
 	this->header.set2(5.0, 1.0, 30.0, 3.0);
 	this->canvas.set(&gs->canvas, 1.0, 5.0, 38.0, 24.0);
 	// tiles in frame
-	this->save1.set2(1.0, 1.0, 36.0, 5.0);
+	this->save1.set("Save 1", 0.33, 1.0, 1.0, 36.0, 5.0);
 }
 
 void fd::FileScreen::event (GameState* gs) {
@@ -22,7 +22,7 @@ void fd::FileScreen::event (GameState* gs) {
 		p.set2(gs->event.mouse.x, gs->event.mouse.y);
 		this->canvas.get_mouse(&p);
 		
-		if (this->save1.contains2(p.x, p.y)) {
+		if (this->save1.contains(&p)) {
 			gs->goto_screen(fd::ScreenWorld);
 			gs->event.consume();
 		}
@@ -34,7 +34,7 @@ void fd::FileScreen::draw (GameState* gs) {
 	this->canvas.clear();
 	gs->graphics.set_color(&gs->draw_color);
 	gs->canvas.draw_rect(&this->canvas.space.rect, ng::DrawFrame);
-	this->canvas.draw_rect(&this->save1, ng::DrawFrame);
+	this->canvas.draw_button(&gs->font, &this->save1);
 	gs->canvas.draw_rect(&this->header, ng::DrawFrame);
 }